+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/service/dto/MenuRequest.java
@@ -1,4 +1,4 @@
-package com.jjangiji.hankkimoa.restaurant.service;
+package com.jjangiji.hankkimoa.restaurant.service.dto;
 
 public record MenuRequest(boolean isMain, String name,
                           String introduce, Integer price, String imgUrl) {
