+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/service/dto/OpeningHoursDetailRequest.java
@@ -1,4 +1,4 @@
-package com.jjangiji.hankkimoa.restaurant.service;
+package com.jjangiji.hankkimoa.restaurant.service.dto;
 
 import java.time.LocalTime;
