+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/service/dto/OpeningHoursRequest.java
@@ -1,4 +1,4 @@
-package com.jjangiji.hankkimoa.restaurant.service;
+package com.jjangiji.hankkimoa.restaurant.service.dto;
 
 public record OpeningHoursRequest(String dayOfWeek, OpeningHoursDetailRequest hours) {
