+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/restaurant/service/dto/RestaurantCreateRequest.java
@@ -1,0 +1,8 @@
+package com.jjangiji.hankkimoa.restaurant.service.dto;
+
+import java.util.List;
+
+public record RestaurantCreateRequest(String id, String name, String category, String address, List<String> images,
+                                      List<OpeningHoursRequest> openingHours,
+                                      List<MenuRequest> menus) {
+}
