--- backend/hankkimoa/src/test/java/com/jjangiji/hankkimoa/restaurant/service/RestaurantCreateRequest.java
@@ -1,8 +1,0 @@
-package com.jjangiji.hankkimoa.restaurant.service;
-
-import java.util.List;
-
-public record RestaurantCreateRequest(String id, String name, String category, String address, List<String> images,
-                                      OpeningHoursRequest openingHours,
-                                      MenuRequest menu) {
-}
